% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{count_dups}
\alias{count_dups}
\title{count_dups}
\usage{
count_dups(df, ..., sort = FALSE)
}
\arguments{
\item{df}{The source dataframe}

\item{...}{A select style set of variables to check for duplicates.}

\item{sort}{whether to sort the results by the number of duplicates. Default is FALSE}
}
\description{
Returns a count of duplicate values for each duplicated value.
That is, if you have a column that is supposed to be unique,
but there are two instances of a given value,
it will return a count of 2 for that value.
}
\examples{
count_dups(iris, Sepal.Width, Sepal.Length)
}
\keyword{duplicates}
