% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal_nrow.R
\name{equal_nrow}
\alias{equal_nrow}
\title{equal_nrow}
\usage{
equal_nrow(x, y)
}
\arguments{
\item{x}{First object to check}

\item{y}{Second object to check}
}
\description{
Check if two objects have equal number of rows.
If there are differences, print the sizes.
}
\examples{
x <- 1:10
y <- 1:9
equal_nrow(x, y)
}
