% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{get_dups}
\alias{get_dups}
\title{get_dups}
\usage{
get_dups(df, ...)
}
\arguments{
\item{df}{The source dataframe}

\item{...}{A select style set of variables to check for duplicates.}
}
\description{
Return all instances of duplicated values from a dataframe.
That is, if a column is supposed to be unique, but there are two identical
values, it will filter the dataframe to both instances of that value.
}
\examples{
sep_dups(iris, "Sepal.Width", "Sepal.Length")
}
\keyword{duplicates}
