% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_id.R
\name{is_id}
\alias{is_id}
\title{is_id}
\usage{
is_id(x)
}
\arguments{
\item{x}{potential unique identifier}
}
\description{
Check if observations are uniquely identified.
If there are duplicates, print the number of duplicates.
}
\examples{
id <- 1:10
is_id(id)
}
\keyword{duplicates,}
\keyword{id,}
\keyword{is_id}

