% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_folder.R
\name{read_csv_folder}
\alias{read_csv_folder}
\title{read_csv_folder}
\usage{
read_csv_folder(path, pattern = ".csv", col_types = NULL)
}
\arguments{
\item{path}{path to the folder you want to read}

\item{pattern}{passthrough pattern for list.files}

\item{col_types}{passthrough col_types to read_csv}
}
\description{
Read csv for each file in a folder.
Useful if you have a datasource that was broken up into multiple files,
e.g., one file per year, and you want to combine the files into a single result.
Returns a list of tibbles (you will do the combining on your own).
}
