% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dups.R
\name{sep_dups}
\alias{sep_dups}
\title{sep_dups}
\usage{
sep_dups(df, ...)
}
\arguments{
\item{df}{The source dataframe}

\item{...}{A select style set of variables to check for duplicates. Uses standard evaluation, so you need to wrap in quotes}
}
\description{
Return a list of dataframes two elements
}
\details{
* .$unique - observations that have no duplicate elements
(i.e., it does not pick one from the duplicated elements)
* .$duplicates - observations that have any duplicate elements
(i.e., if there are two duplicates for an observation, it returns both of them)

The `length(unique)` + `length(duplicates)` should equal the length of df
}
\examples{
sep_dups(iris, "Sepal.Width", "Sepal.Length")
}
\keyword{duplicates}
