% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete.R
\name{is_complete}
\alias{is_complete}
\title{is_complete}
\usage{
is_complete(x)
}
\arguments{
\item{x}{The thing to check for completeness}
}
\description{
Check if observations are never NULL or NA.
}
\examples{
x <- 1:10
is_complete(x)
}
